fix vis loss

+++ projects/mmdet3d_plugin/models/detection3d/detection3d_head.py
@@ -689,7 +689,7 @@ class Sparse4DHead(BaseModule):
                 matched = mask_valid.reshape(-1)
                 vis_loss = self.loss_visibility(
                     vis.squeeze(-1).flatten(end_dim=1)[matched],
-                    vis_target.flatten(end_dim=1)[matched],
+                    vis_target.flatten(end_dim=1)[matched].long(),
                     avg_factor=num_pos,
                 )
                 output[
